--- a/esphome/components/api/__init__.py
+++ b/esphome/components/api/__init__.py
@@ -177,7 +177,11 @@ async def to_code(config):
             # and plaintext disabled. Only a factory reset can remove it.
             cg.add_define("USE_API_PLAINTEXT")
         cg.add_define("USE_API_NOISE")
-        cg.add_library("esphome/noise-c", "0.1.6")
+        cg.add_library(
+            None,
+            None,
+            "https://github.com/esphome/noise-c.git#libsodium_update",
+        )
     else:
         cg.add_define("USE_API_PLAINTEXT")