--- a/esphome/components/api/__init__.py
+++ b/esphome/components/api/__init__.py
@@ -182,7 +182,7 @@ async def to_code(config):
             # and plaintext disabled. Only a factory reset can remove it.
             cg.add_define("USE_API_PLAINTEXT")
         cg.add_define("USE_API_NOISE")
-        cg.add_library("esphome/noise-c", "0.1.8")
+        cg.add_library("esphome/noise-c", "0.1.9")
     else:
         cg.add_define("USE_API_PLAINTEXT")