--- a/platformio.ini
+++ b/platformio.ini
@@ -33,7 +33,7 @@ build_flags =
 ; This are common settings for all environments.
 [common]
 lib_deps =
-    esphome/noise-c@0.1.8                  ; api
+    esphome/noise-c@0.1.9                  ; api
     improv/Improv@1.2.4                    ; improv_serial / esp32_improv
     bblanchon/ArduinoJson@6.18.5           ; json
     wjtje/qr-code-generator-library@1.7.0  ; qr_code
@@ -556,7 +556,7 @@ build_unflags =
 extends = common
 platform = platformio/native
 lib_deps =
-    esphome/noise-c@0.1.8  ; used by api
+    esphome/noise-c@0.1.9  ; used by api
 build_flags =
     ${common.build_flags}
     -DUSE_HOST