# What does this implement/fix?

Updates libsodium from 1.0.18 to 1.0.20, bringing security fixes, bug fixes, and performance improvements. The handshake time for noise-encrypted connections improved from 0.70s to 0.64s (~9% faster).

This update also fixes the build configuration to include only the cryptographic primitives actually used by ESPHome's noise-c implementation, maintaining nearly the same flash footprint while gaining the benefits of the newer version.

## Types of changes

- [x] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [x] Code quality improvements to existing code or addition of tests
- [ ] Other

**Related issue or feature (if applicable):**

- Updates libsodium to address security vulnerabilities and bugs fixed between 1.0.18 and 1.0.20

**Pull request in [esphome-docs](https://github.com/esphome/esphome-docs) with documentation (if applicable):**

- N/A (no user-facing changes)

## Test Environment

- [x] ESP32
- [x] ESP32 IDF
- [ ] ESP8266
- [ ] RP2040
- [ ] BK72xx
- [ ] RTL87xx

## Example entry for `config.yaml`:

```yaml
# No configuration changes needed - this is an internal dependency update
api:
  encryption:
    key: "your-encryption-key-here"
```

## Checklist:
  - [x] The code change is tested and works locally.
  - [ ] Tests have been added to verify that the new code works (under `tests/` folder).

If user exposed functionality or configuration variables are added/changed:
  - [ ] Documentation added/updated in [esphome-docs](https://github.com/esphome/esphome-docs).

## Additional Notes

Changes made:
1. Updated libsodium submodule from 1.0.18 to 1.0.20
2. Fixed `library.json` to include `crypto_verify/verify.c` (was incorrectly specified as `crypto_verify/sodium/verify.c`)
3. Optimized build configuration to include only components used by noise-c:
   - Kept: Curve25519, ChaCha20-Poly1305, SHA256, and required utilities
   - Removed unused components: Blake2b, SHA512, Ed25519 signing, Argon2, Ristretto255, AEGIS AEAD ciphers

Flash usage comparison:
- Before: 1,063,728 bytes (58.0%)
- After: 1,063,640 bytes (58.0%)
- Difference: -88 bytes (slight reduction)

Performance improvement:
- Handshake time reduced from 0.70s to 0.64s