@bdraco 2025-06-27
Update looks to be 9% faster