@bdraco 2025-06-27
cleaned up the libsodium files a bit, reduce flash by 88 bytes... its something