@bdraco 2025-06-27
I did test publishes to pio registry under my personal account