@bdraco 2025-06-30
It looks like the patch isn't being applied somehow

```
/usr/bin/ld: .pioenvs/host-noise-test/lib8fc/libsodium.a(core.o): in function `sodium_init':
/tmp/tmpdhas4c_b/.esphome/build/host-noise-test/.piolibdeps/host-noise-test/libsodium/libsodium/src/libsodium/sodium/core.c:42:(.text+0x49): undefined reference to `_crypto_pwhash_argon2_pick_best_implementation'
/usr/bin/ld: /tmp/tmpdhas4c_b/.esphome/build/host-noise-test/.piolibdeps/host-noise-test/libsodium/libsodium/src/libsodium/sodium/core.c:43:(.text+0x4e): undefined reference to `_crypto_generichash_blake2b_pick_best_implementation'
/usr/bin/ld: /tmp/tmpdhas4c_b/.esphome/build/host-noise-test/.piolibdeps/host-noise-test/libsodium/libsodium/src/libsodium/sodium/core.c:48:(.text+0x67): undefined reference to `_crypto_aead_aegis128l_pick_best_implementation'
/usr/bin/ld: /tmp/tmpdhas4c_b/.esphome/build/host-noise-test/.piolibdeps/host-noise-test/libsodium/libsodium/src/libsodium/sodium/core.c:49:(.text+0x6c): undefined reference to `_crypto_aead_aegis256_pick_best_implementation'
```